+++ tests/walk_controller_test.cpp
@@ -18,7 +18,7 @@ int main() {
     assert(wc.setGaitType(TRIPOD_GAIT));
     wc.updateGaitPhase(0.5f);
     LegState states[NUM_LEGS]{};
-    float offsets[NUM_LEGS]{};
+    double offsets[NUM_LEGS]{};
     Point3D traj = wc.footTrajectory(0, 0.2f, 20, 50, 0.5f, 0.5f, p.robot_height, offsets, states, nullptr, nullptr);
     (void)traj;
     std::cout << "walk_controller_test executed successfully" << std::endl;
